Move IndexedBlueprintSchema
