+++ transaction/src/manifest/e2e.rs
@@ -24,7 +24,10 @@ PUBLISH_PACKAGE_ADVANCED
     Enum<0u8>()
     Blob("${code_blob_hash}")
     Tuple(
-        Map<String, Tuple>()
+        Tuple(
+            Map<String, Tuple>()
+        ),
+        Map<String, Map>()
     )
     Map<String, Tuple>()
     Map<String, Enum>()
